+++ packages/dds/tree/README.md
@@ -302,7 +302,7 @@ flowchart
         direction TB
         subgraph core ["core libraries"]
             direction TB
-            checkout-->rebase
+            checkout
             forest-->schema-stored
             change-family-->repair
             edit-manager-->change-family
