+++ packages/dds/tree/src/core/checkout/fence.json
@@ -1,5 +1,5 @@
 {
 	"tags": ["checkout"],
 	"exports": ["index"],
-	"imports": ["events", "rebase"]
+	"imports": []
 }
