+++ packages/dds/tree/src/core/edit-manager/fence.json
@@ -1,5 +1,16 @@
 {
-	"tags": ["edit-manager"],
-	"exports": ["index"],
-	"imports": ["change-family", "dependency-tracking", "rebase", "tree", "util"]
+	"tags": [
+		"edit-manager"
+	],
+	"exports": [
+		"index"
+	],
+	"imports": [
+		"change-family",
+		"dependency-tracking",
+		"rebase",
+		"repair",
+		"tree",
+		"util"
+	]
 }
