+++ packages/dds/tree/src/core/index.ts
@@ -151,16 +151,12 @@ export {
 	verifyChangeRebaser,
 	tagInverse,
 	SessionId,
+	mintCommit,
 	mintRevisionTag,
 	Rebaser,
 } from "./rebase";
 
-export {
-	SharedTreeBranch,
-	SharedTreeBranchEvents,
-	TransactionResult,
-	TransactionStack,
-} from "./checkout";
+export { TransactionResult } from "./checkout";
 
 export { Checkout as TransactionCheckout } from "./transaction";
