+++ packages/dds/tree/src/shared-tree-core/fence.json
@@ -1,5 +1,5 @@
 {
 	"tags": ["shared-tree-core"],
 	"exports": ["index"],
-	"imports": ["core", "events", "util"]
+	"imports": ["core", "events", "util", "rebase"]
 }
