+++ packages/dds/tree/src/shared-tree-core/index.ts
@@ -2,6 +2,7 @@
  * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
  * Licensed under the MIT License.
  */
+export { SharedTreeBranch, SharedTreeBranchEvents } from "./branch";
 
 export {
 	Index,
@@ -12,3 +13,5 @@ export {
 	SummaryElementParser,
 	SummaryElementStringifier,
 } from "./sharedTreeCore";
+
+export { TransactionStack } from "./transactionStack";
